feat(terms): WIP
