+++ core/bootstrap/bootstrap-constructor-mock.ts
@@ -2,6 +2,7 @@ import { BootstrapConstructor } from "core/bootstrap/index";
 import { ContactAdapterMock } from "core/helpers/contact/contact-adapter-mock";
 import { DateAdapterMock } from "core/helpers/date/date-adapter-mock";
 import { ImageAdapterMock } from "core/helpers/image/image-adapter-mock";
+import { LegalAdapterMock } from "core/helpers/legal/legal-adapter-mock";
 import { UrlAdapterMock } from "core/helpers/url/url-adapter-mock";
 import { HackathonClientAdapterMock } from "core/infrastructure/marketplace-api-client-adapter/mock-adapters/hackathon-client-adapter-mock";
 import { ProjectClientAdapterMock } from "core/infrastructure/marketplace-api-client-adapter/mock-adapters/project-client-adapter-mock";
@@ -18,4 +19,5 @@ export const bootstrapConstructorMock: BootstrapConstructor = {
   urlHelperPort: UrlAdapterMock,
   imageHelperPort: new ImageAdapterMock(),
   contactHelperPort: new ContactAdapterMock(),
+  legalHelperPort: new LegalAdapterMock(),
 };
