+++ core/bootstrap/index.ts
@@ -7,6 +7,8 @@ import { DateFacadePort } from "core/helpers/date/date-facade-port";
 import { DateFnsAdapter } from "core/helpers/date/date-fns-adapter";
 import { ImageAdapter } from "core/helpers/image/image-adapter";
 import { ImageFacadePort } from "core/helpers/image/image-facade-port";
+import { LegalAdapter } from "core/helpers/legal/legal-adapter";
+import { LegalFacadePort } from "core/helpers/legal/legal-facade-port";
 import { UrlAdapter } from "core/helpers/url/url-adapter";
 import { UrlFacadePort } from "core/helpers/url/url-facade-port";
 import { HackathonClientAdapter } from "core/infrastructure/marketplace-api-client-adapter/adapters/hackathon-client-adapter";
@@ -27,6 +29,7 @@ export interface BootstrapConstructor {
   urlHelperPort: UrlFacadePort;
   imageHelperPort: ImageFacadePort;
   contactHelperPort: ContactFacadePort;
+  legalHelperPort: LegalFacadePort;
 }
 
 export class Bootstrap {
@@ -43,6 +46,7 @@ export class Bootstrap {
   urlHelperPort: UrlFacadePort;
   imageHelperPort: ImageFacadePort;
   contactHelperPort: ContactFacadePort;
+  legalHelperPort: LegalFacadePort;
 
   constructor(constructor: BootstrapConstructor) {
     this.projectStoragePortForClient = constructor.projectStoragePortForClient;
@@ -55,6 +59,7 @@ export class Bootstrap {
     this.urlHelperPort = constructor.urlHelperPort;
     this.imageHelperPort = constructor.imageHelperPort;
     this.contactHelperPort = constructor.contactHelperPort;
+    this.legalHelperPort = constructor.legalHelperPort;
   }
 
   getAuthProvider() {
@@ -113,6 +118,10 @@ export class Bootstrap {
     return this.contactHelperPort;
   }
 
+  getLegalHelperPort() {
+    return this.legalHelperPort;
+  }
+
   public static get getBootstrap(): Bootstrap {
     if (!Bootstrap.#instance) {
       this.newBootstrap({
@@ -126,6 +135,7 @@ export class Bootstrap {
         urlHelperPort: UrlAdapter,
         imageHelperPort: new ImageAdapter(),
         contactHelperPort: new ContactAdapter(),
+        legalHelperPort: new LegalAdapter(),
       });
     }
