+++ core/helpers/legal/legal-facade-port.ts
@@ -1,0 +1,6 @@
+export interface LegalFacadePort {
+  urls: {
+    terms: string;
+    privacy: string;
+  };
+}
