+++ translations/v2/en/pages/signup/onboarding/onboarding-terms.json
@@ -11,8 +11,7 @@
       "description": "To use our platform, you must be of legal age (+ 16 years), not be a sanctioned person, and, of course, accept and observe the T&Cs in full. To obtain your rewards, you will need to comply with KYC/KYB requirements."
     },
     "responsability": {
-      "title": "What are your responsibilities?",
-      "description": "You can use the platform as a project leader or as a contributor.\n\n__If you are a contributor__, you will provide contributions from among those proposed on the platform by the projects. You undertake to execute these contributions with the highest level of diligence, in accordance with the specifications set out by the project. Rewards for this contribution will not be systematic and will depend in part on how well the contribution is executed. This evaluation will be carried out by the project leader and, except in cases of fraud, Only Dust will not be directly involved.\n\n__If you are a Project Leader__, the project must be open-source and useful for the community. You will be responsible for the project, and as such, you will have to define the nature of the contributions required, assess their quality and determine the reward amount to be paid to contributors. You can receive a grant depending on the merits of your project, which is purely discretionary (will be taken by OD or by a committee of independent experts)\n\n__In both cases__, you must: (i) properly use our platform and refrain from all fraudulent activity; (ii) give us valid information about your status, including whether you act for a company or not; (iii) comply with the laws and regulations in force in the country in which you are located.\n\n**What are Our responsibilities?** We undertake to provide you with our platform and ensure its proper operation and transfer to contributors the rewards under the conditions defined by the project leader.\n\nBy using Our platform, You understand that we are not responsible for the interruption or breakdowns of our platform; we are not your employer and the rewards that we transfer to you must not be considered as a salary; we are not responsible for the amount of budget that is given to a project nor for the amount of rewards that is given to contributors; and we are not responsible if the foundations decide not to give us funds anymore."
+      "title": "What are your responsibilities?"
     }
   },
   "agree": {
